Refactor narwhal into subcrates for block usage
