+++ Cargo.toml
@@ -66,6 +66,10 @@ members = [
   "ledger/block",
   "ledger/coinbase",
   "ledger/narwhal",
+  "ledger/narwhal/batch-certificate",
+  "ledger/narwhal/batch-header",
+  "ledger/narwhal/transmission",
+  "ledger/narwhal/transmission-id",
   "ledger/query",
   "ledger/store",
   "ledger/test-helpers",
