+++ ledger/narwhal/Cargo.toml
@@ -25,60 +25,53 @@ edition = "2021"
 
 [features]
 default = [ "batch-certificate", "batch-header", "transmission", "transmission-id" ]
-serial = [ "console/serial" ]
-wasm = [ "console/wasm" ]
-test-helpers = [ "bytes", "time" ]
+serial = [
+    "narwhal-batch-certificate/serial",
+    "narwhal-batch-header/serial",
+    "narwhal-transmission/serial",
+    "narwhal-transmission-id/serial",
+]
+wasm = [
+    "narwhal-batch-certificate/wasm",
+    "narwhal-batch-header/wasm",
+    "narwhal-transmission/wasm",
+    "narwhal-transmission-id/wasm",
+]
+test-helpers = [
+    "narwhal-batch-certificate/test-helpers",
+    "narwhal-batch-header/test-helpers",
+    "narwhal-transmission/test-helpers",
+    "narwhal-transmission-id/test-helpers",
+]
 
-batch-certificate = [ "batch-header", "console", "indexmap", "serde_json", "transmission-id" ]
-batch-header = [ "console", "indexmap", "transmission-id" ]
-transmission = [ "bytes", "ledger-block", "ledger-coinbase" ]
-transmission-id = [ "console", "ledger-coinbase" ]
+batch-certificate = [ "narwhal-batch-certificate" ]
+batch-header = [ "narwhal-batch-header" ]
+transmission = [ "narwhal-transmission" ]
+transmission-id = [ "narwhal-transmission-id" ]
 
-[dependencies.console]
-package = "snarkvm-console"
-path = "../../console"
+[dependencies.narwhal-batch-certificate]
+package = "snarkvm-ledger-narwhal-batch-certificate"
+path = "./batch-certificate"
 version = "=0.14.5"
 optional = true
 
-[dependencies.ledger-block]
-package = "snarkvm-ledger-block"
-path = "../../ledger/block"
+[dependencies.narwhal-batch-header]
+package = "snarkvm-ledger-narwhal-batch-header"
+path = "./batch-header"
 version = "=0.14.5"
 optional = true
 
-[dependencies.ledger-coinbase]
-package = "snarkvm-ledger-coinbase"
-path = "../../ledger/coinbase"
+[dependencies.narwhal-transmission]
+package = "snarkvm-ledger-narwhal-transmission"
+path = "./transmission"
 version = "=0.14.5"
 optional = true
 
-[dependencies.bytes]
-version = "1"
+[dependencies.narwhal-transmission-id]
+package = "snarkvm-ledger-narwhal-transmission-id"
+path = "./transmission-id"
+version = "=0.14.5"
 optional = true
-
-[dependencies.indexmap]
-version = "2.0"
-features = [ "serde" ]
-optional = true
-
-[dependencies.serde_json]
-version = "1.0"
-optional = true
-features = [ "preserve_order" ]
-
-[dependencies.time]
-version = "0.3"
-optional = true
-
-[dev-dependencies.bincode]
-version = "1.3"
-
-[dev-dependencies.rand]
-version = "0.8"
-
-[dev-dependencies.serde_json]
-version = "1.0"
-features = [ "preserve_order" ]
 
 [dev-dependencies.snarkvm-ledger-narwhal]
 path = "."
