+++ ledger/narwhal/batch-certificate/Cargo.toml
@@ -1,0 +1,68 @@
+[package]
+name = "snarkvm-ledger-narwhal-batch-certificate"
+version = "0.14.5"
+authors = [ "The Aleo Team <hello@aleo.org>" ]
+description = "A batch certificate for a Narwhal-style memory pool in a decentralized virtual machine"
+homepage = "https://aleo.org"
+repository = "https://github.com/AleoHQ/snarkVM"
+keywords = [
+    "aleo",
+    "cryptography",
+    "blockchain",
+    "decentralized",
+    "zero-knowledge"
+]
+categories = [
+    "compilers",
+    "cryptography",
+    "mathematics",
+    "wasm",
+    "web-programming"
+]
+include = [ "Cargo.toml", "src", "README.md", "LICENSE.md" ]
+license = "Apache-2.0"
+edition = "2021"
+
+[features]
+default = [ ]
+serial = [ "console/serial" ]
+wasm = [ "console/wasm" ]
+test-helpers = [ "narwhal-batch-header/test-helpers", "time" ]
+
+[dependencies.console]
+package = "snarkvm-console"
+path = "../../../console"
+version = "=0.14.5"
+
+[dependencies.narwhal-batch-header]
+package = "snarkvm-ledger-narwhal-batch-header"
+path = "../batch-header"
+version = "=0.14.5"
+
+[dependencies.narwhal-transmission-id]
+package = "snarkvm-ledger-narwhal-transmission-id"
+path = "../transmission-id"
+version = "=0.14.5"
+
+[dependencies.indexmap]
+version = "2.0"
+features = [ "serde" ]
+
+[dependencies.serde_json]
+version = "1.0"
+features = [ "preserve_order" ]
+
+[dependencies.time]
+version = "0.3"
+optional = true
+
+[dev-dependencies.bincode]
+version = "1.3"
+
+[dev-dependencies.serde_json]
+version = "1.0"
+features = [ "preserve_order" ]
+
+[dev-dependencies.snarkvm-ledger-narwhal-batch-certificate]
+path = "."
+features = [ "test-helpers" ]
