+++ ledger/narwhal/batch-certificate/README.md
@@ -1,0 +1,7 @@
+# snarkvm-ledger-narwhal-batch-certificate
+
+[![Crates.io](https://img.shields.io/crates/v/snarkvm-ledger-narwhal-batch-certificate.svg?color=neon)](https://crates.io/crates/snarkvm-ledger-narwhal-batch-certificate)
+[![Authors](https://img.shields.io/badge/authors-Aleo-orange.svg)](https://aleo.org)
+[![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](./LICENSE.md)
+
+The `snarkvm-ledger-narwhal-batch-certificate` crate provides a batch certificate for a Narwhal-style memory pool.
