+++ ledger/narwhal/batch-certificate/src/serialize.rs
@@ -14,22 +14,36 @@
 
 use super::*;
 
-impl<N: Network> Serialize for TransmissionID<N> {
-    #[inline]
+impl<N: Network> Serialize for BatchCertificate<N> {
+    /// Serializes the batch certificate to a JSON-string or buffer.
     fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
         match serializer.is_human_readable() {
-            true => serializer.collect_str(self),
+            true => {
+                let mut certificate = serializer.serialize_struct("BatchCertificate", 3)?;
+                certificate.serialize_field("certificate_id", &self.certificate_id)?;
+                certificate.serialize_field("batch_header", &self.batch_header)?;
+                certificate.serialize_field("signatures", &self.signatures)?;
+                certificate.end()
+            }
             false => ToBytesSerializer::serialize_with_size_encoding(self, serializer),
         }
     }
 }
 
-impl<'de, N: Network> Deserialize<'de> for TransmissionID<N> {
-    #[inline]
+impl<'de, N: Network> Deserialize<'de> for BatchCertificate<N> {
+    /// Deserializes the batch certificate from a JSON-string or buffer.
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
         match deserializer.is_human_readable() {
-            true => FromStr::from_str(&String::deserialize(deserializer)?).map_err(de::Error::custom),
-            false => FromBytesDeserializer::<Self>::deserialize_with_size_encoding(deserializer, "transmission ID"),
+            true => {
+                let mut value = serde_json::Value::deserialize(deserializer)?;
+                Ok(Self::from(
+                    DeserializeExt::take_from_value::<D>(&mut value, "certificate_id")?,
+                    DeserializeExt::take_from_value::<D>(&mut value, "batch_header")?,
+                    DeserializeExt::take_from_value::<D>(&mut value, "signatures")?,
+                )
+                .map_err(de::Error::custom)?)
+            }
+            false => FromBytesDeserializer::<Self>::deserialize_with_size_encoding(deserializer, "batch certificate"),
         }
     }
 }
@@ -48,7 +62,7 @@ mod tests {
         let candidate_string = serde_json::to_string(&expected).unwrap();
         let candidate = serde_json::from_str::<T>(&candidate_string).unwrap();
         assert_eq!(expected, candidate);
-        assert_eq!(expected_string, serde_json::Value::from_str(&candidate_string).unwrap().as_str().unwrap());
+        assert_eq!(expected_string, candidate_string);
         assert_eq!(expected_string, candidate.to_string());
 
         // Deserialize
@@ -75,7 +89,7 @@ mod tests {
     fn test_serde_json() {
         let rng = &mut TestRng::default();
 
-        for expected in crate::transmission_id::test_helpers::sample_transmission_ids(rng) {
+        for expected in crate::test_helpers::sample_batch_certificates(rng) {
             check_serde_json(expected);
         }
     }
@@ -84,7 +98,7 @@ mod tests {
     fn test_bincode() {
         let rng = &mut TestRng::default();
 
-        for expected in crate::transmission_id::test_helpers::sample_transmission_ids(rng) {
+        for expected in crate::test_helpers::sample_batch_certificates(rng) {
             check_bincode(expected);
         }
     }
