+++ ledger/narwhal/batch-certificate/src/string.rs
