+++ ledger/narwhal/batch-header/src/bytes.rs
@@ -109,7 +109,7 @@ mod tests {
     fn test_bytes() {
         let rng = &mut TestRng::default();
 
-        for expected in crate::batch_header::test_helpers::sample_batch_headers(rng) {
+        for expected in crate::test_helpers::sample_batch_headers(rng) {
             // Check the byte representation.
             let expected_bytes = expected.to_bytes_le().unwrap();
             assert_eq!(expected, BatchHeader::read_le(&expected_bytes[..]).unwrap());
