+++ ledger/narwhal/batch-header/src/lib.rs
@@ -17,13 +17,13 @@ mod serialize;
 mod string;
 mod to_id;
 
-use crate::TransmissionID;
 use console::{
     account::{Address, PrivateKey, Signature},
     prelude::*,
     types::Field,
 };
 use indexmap::IndexSet;
+use narwhal_transmission_id::TransmissionID;
 
 #[derive(Clone, PartialEq, Eq)]
 pub struct BatchHeader<N: Network> {
@@ -153,6 +153,7 @@ impl<N: Network> BatchHeader<N> {
 pub mod test_helpers {
     use super::*;
     use console::{account::PrivateKey, network::Testnet3, prelude::TestRng};
+
     use time::OffsetDateTime;
 
     type CurrentNetwork = Testnet3;
@@ -168,7 +169,7 @@ pub mod test_helpers {
         let private_key = PrivateKey::new(rng).unwrap();
         // Sample transmission IDs.
         let transmission_ids =
-            crate::transmission_id::test_helpers::sample_transmission_ids(rng).into_iter().collect::<IndexSet<_>>();
+            narwhal_transmission_id::test_helpers::sample_transmission_ids(rng).into_iter().collect::<IndexSet<_>>();
         // Sample certificate IDs.
         let certificate_ids = (0..10).map(|_| Field::<CurrentNetwork>::rand(rng)).collect::<IndexSet<_>>();
         // Checkpoint the timestamp for the batch.
