+++ ledger/narwhal/transmission-id/src/bytes.rs
@@ -58,7 +58,7 @@ mod tests {
     fn test_bytes() {
         let rng = &mut TestRng::default();
 
-        for expected in crate::transmission_id::test_helpers::sample_transmission_ids(rng) {
+        for expected in crate::test_helpers::sample_transmission_ids(rng) {
             // Check the byte representation.
             let expected_bytes = expected.to_bytes_le().unwrap();
             assert_eq!(expected, TransmissionID::read_le(&expected_bytes[..]).unwrap());
