+++ ledger/narwhal/transmission/Cargo.toml
@@ -1,0 +1,55 @@
+[package]
+name = "snarkvm-ledger-narwhal-transmission"
+version = "0.14.5"
+authors = [ "The Aleo Team <hello@aleo.org>" ]
+description = "A transmission for a Narwhal-style memory pool in a decentralized virtual machine"
+homepage = "https://aleo.org"
+repository = "https://github.com/AleoHQ/snarkVM"
+keywords = [
+    "aleo",
+    "cryptography",
+    "blockchain",
+    "decentralized",
+    "zero-knowledge"
+]
+categories = [
+    "compilers",
+    "cryptography",
+    "mathematics",
+    "wasm",
+    "web-programming"
+]
+include = [ "Cargo.toml", "src", "README.md", "LICENSE.md" ]
+license = "Apache-2.0"
+edition = "2021"
+
+[features]
+default = [ ]
+serial = [ "console/serial" ]
+wasm = [ "console/wasm" ]
+test-helpers = [ ]
+
+[dependencies.console]
+package = "snarkvm-console"
+path = "../../../console"
+version = "=0.14.5"
+
+[dependencies.ledger-block]
+package = "snarkvm-ledger-block"
+path = "../../block"
+version = "=0.14.5"
+
+[dependencies.ledger-coinbase]
+package = "snarkvm-ledger-coinbase"
+path = "../../coinbase"
+version = "=0.14.5"
+
+[dependencies.bytes]
+version = "1"
+
+[dependencies.serde_json]
+version = "1.0"
+features = [ "preserve_order" ]
+
+[dev-dependencies.bincode]
+version = "1.3"
